Update Header 1.18.30.04
PDB_Hash b1918f64d4baed89f791c5e23c557aee17f8f1d43d12d5ab395f2f0aa47cefcf
EXE_Hash 10723c5824d2e70c707d582dd2eb4cb55da147225de6503e9cbde20be944361a

+++ LiteLoader/Header/MC/AttackDamageMobEffect.hpp
@@ -18,17 +18,17 @@ class AttackDamageMobEffect : public MobEffect {
 
 #ifndef DISABLE_CONSTRUCTOR_PREVENTION_ATTACKDAMAGEMOBEFFECT
 public:
-    class AttackDamageMobEffect& operator=(class AttackDamageMobEffect const&) = delete;
-    AttackDamageMobEffect(class AttackDamageMobEffect const&) = delete;
+    class AttackDamageMobEffect& operator=(class AttackDamageMobEffect const &) = delete;
+    AttackDamageMobEffect(class AttackDamageMobEffect const &) = delete;
     AttackDamageMobEffect() = delete;
 #endif
 
 public:
     /*0*/ virtual ~AttackDamageMobEffect();
-    /*5*/ virtual float getAttributeModifierValue(int, class AttributeModifier const&) const;
-    MCAPI AttackDamageMobEffect(int, std::string const&, std::string const&, bool, int, int);
-    MCAPI static int getDamageAfterDamageBoost(int, int);
-    MCAPI static int getDamageAfterWeakness(int, int);
+    /*5*/ virtual float getAttributeModifierValue(int, class AttributeModifier const &) const;
+    MCAPI AttackDamageMobEffect(int, std::string const &, std::string const &, bool, int, int);
+    MCAPI static float getDamageAfterDamageBoost(float, int);
+    MCAPI static float getDamageAfterWeakness(float, int);
 
 protected:
